+++ libnopegl/src/test_colorconv.c
@@ -89,7 +89,7 @@ static int compare_matrices(const float *a, const float *b)
 {
     int fail = 0;
     float diff[4 * 4];
-    for (int i = 0; i < NGLI_ARRAY_NB(diff); i++) {
+    for (size_t i = 0; i < NGLI_ARRAY_NB(diff); i++) {
         diff[i] = fabsf(a[i] - b[i]);
         fail += diff[i] > 1e-6;
     }
@@ -103,9 +103,9 @@ int main(void)
     float mat[4 * 4];
     struct color_info cinfo = NGLI_COLOR_INFO_DEFAULTS;
 
-    for (int r = 0; r < NGLI_ARRAY_NB(ranges); r++) {
+    for (size_t r = 0; r < NGLI_ARRAY_NB(ranges); r++) {
         cinfo.range = ranges[r].val;
-        for (int s = 0; s < NGLI_ARRAY_NB(spaces); s++) {
+        for (size_t s = 0; s < NGLI_ARRAY_NB(spaces); s++) {
             cinfo.space = spaces[s].val;
             if (ngli_colorconv_get_ycbcr_to_rgb_color_matrix(mat, &cinfo, 1.f) < 0)
                 return 1;
